requirements.txt
streamlit
numpy 
matplotlib.pyplot
tensorflow-cpu
requests